+++ src/dash-table/reconcile/isNully.ts
@@ -1,0 +1,1 @@
+export default (value: any) => value === undefined || value === null || (typeof value === 'number' && isNaN(value));