+++ src/dash-table/reconcile/text.ts
@@ -1,4 +1,5 @@
 import { ITextColumn } from 'dash-table/components/Table/props';
+import isNully from './isNully';
 
 export function coerce(value: any, options?: ITextColumn) {
     const validation = validate(value, options);
@@ -6,22 +7,24 @@ export function coerce(value: any, options?: ITextColumn) {
         return validation;
     }
 
-    const allowNully = Boolean(options && options.validation && options.validation.allow_nully);
-    const isNully = value === undefined || value === null;
+    const allowNull = Boolean(options && options.validation && options.validation.allow_null);
+    const nully = isNully(value);
 
     return {
-        success: Boolean(!isNully || allowNully),
-        value: isNully ? value : JSON.stringify(value)
+        success: Boolean(!nully || allowNull),
+        value: nully ? value : JSON.stringify(value)
     };
 }
 
 export function validate(value: any, options?: ITextColumn) {
-    const allowNully = Boolean(options && options.validation && options.validation.allow_nully);
+    const allowNull = Boolean(options && options.validation && options.validation.allow_null);
 
-    const isNully = value === undefined || value === null;
-    if (isNully && allowNully) {
-        return { success: true, value };
+    const nully = isNully(value);
+    if (nully && allowNull) {
+        return { success: true, value: null };
     }
+
+    value = nully ? null : value;
 
     if (typeof value === 'string') {
         return { success: true, value };